ids-draggable PR feedback (attrib, example html, util rename)
