+++ src/ids-base/ids-attributes.js
@@ -23,8 +23,6 @@ export const attributes = {
   BLEED: 'bleed',
   BUTTONS: 'buttons',
   CANCEL: 'cancel',
-  BOUNDED_BY_PARENT: 'bounded-by-parent',
-  PARENT_CONTAINMENT: 'parent-containment',
   CHAR_MAX_TEXT: 'char-max-text',
   CHAR_REMAINING_TEXT: 'char-remaining-text',
   CHARACTER_COUNTER: 'character-counter',
@@ -112,6 +110,7 @@ export const attributes = {
   PAGE_NUMBER: 'page-number',
   PAGE_SIZE: 'page-size',
   PADDING: 'padding',
+  PARENT_CONTAINMENT: 'parent-containment',
   PARENT_DISABLED: 'parent-disabled',
   PERCENTAGE_VISIBLE: 'percentage-visible',
   PLACEHOLDER: 'placeholder',
